change: Moved some stuff over from Orbis into OrbisAPI. feature: Many GuiVar implementations for Blueprint GUIS. feature/change: FilterOptions now use GuiVars. change: Schedule layers now use node/tree data structure for link relationships. change/feature: BakedBlueprints now resolve conditions in schedule layers and only use the baked schedule layers that have all of its conditions resolved. change: Rect/Dim2D implementation now uses Pos2D objects for min and max instead of primitive values. Might change if memory/object creation impact is high. feature: Many utility GUIs for Orbis. change: BlockFilterHelper returns a layer for getNewFillLayer() that only places in air blocks.

+++ src/main/java/com/gildedgames/orbis_api/client/gui/util/GuiInputSlider.java
@@ -24,7 +24,7 @@ public class GuiInputSlider extends GuiFrame
 
 	private DecimalFormat df;
 
-	public GuiInputSlider(final Rect rect, int minValue, int maxValue, float sliderValue)
+	public GuiInputSlider(final Rect rect, float minValue, float maxValue, float sliderValue)
 	{
 		super(rect);
